requirements.txt
alembic==1.16.4
annotated-types==0.7.0
anyio==4.9.0
boto3==1.34.162
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
Mako==1.3.10
MarkupSafe==3.0.2
psycopg2-binary==2.9.10
pydantic==2.11.7
pydantic_core==2.33.2
python-dotenv==1.1.1
sniffio==1.3.1
SQLAlchemy==2.0.41
starlette==0.47.1
typing-inspection==0.4.1
typing_extensions==4.14.1
uvicorn==0.35.0
